Feat: P156 - Refactoring about enemy controller
